requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-tinymce==3.3.0
django-tinymce-lite==0.0.2
django-tinymce4-lite==1.8.0
djangorestframework==3.12.4
idna==2.10
jsmin==2.2.2
oauthlib==3.1.0
Pillow==8.2.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
